clickbot_archive/clickbot_v2/requirements.txt
PyQt5==5.15.11
pyautogui==0.9.54
opencv-python>=4.5.0
numpy>=1.19.0
pytesseract>=0.3.8
Pillow>=8.0.0
mss>=6.1.0 